requirements.txt
flask
pillow
werkzeug
ultralytics
opencv-python
numpy